Resolved jquery bug
The bug from jquery in the graphic page was resolved. The problem was that inside 'application.js' the directory for jquery was mistyped. For that not just the javascript directory files were changed in the 'application.js', but also the 'css' files were putted to precompile in the asset pipeline for no risk of bugs and for better performance. Finally, the 'html.erb' files were changed to remove the javascript call and to add a stylesheet_tag.