Spreadsheet search added in menu bar
The search bar can now find spreadsheets across the database acording to the selected month and them redirect the user to the previus page with the updated data. Also some changes have been made in the table Report, changing the column type month to string and adding a new column called month_numb which has the month number representation, and in the route.tb file, adding the paths to search the spreadsheet depending on the redirect page.