Restructure of controllers and new report page added
The controller structure have been modified to concentrate specific methods. With that, the Dashboard controls the webpages, the Monthly is responsable to the searchs of specific months and the Overview focus in the overall data search. Also, the webpages have been modified to merge the graphic and spreadsheet into monthly_sales and create a new page called report_page which will show the sales record for every salesman in a certain month.