Overview report added with code style review
A new webpage called overview_reports have been added to compare all the salesman in a period of months, similar to overview. For that to happend, the methods and webpages have been changed to two different names: overview_months to compare the report goals; overview_report to compare the salesman. Also, the other files listed have been modified in the code style with rubocop.