New language options for the user
The application support two languages for the users, portuguese and english. New files in locale were added about the portuguese text for the normal pages and devise gem. The english locale have been updated for the new text. For users who access the system in the previous links without a locale set, a new page have been created to redirect the user for the official page with a defaoult locale.