Remove Action Cable from the system
In the 18th PR, the action cable were discovered to not been used inside the application in any environments. So the cable.js, the channels folder, the action cable requirements, the cable.yml in the configs and the comments inside product environment have been removed with also updating the gemfiles.