Add helpers for Overview, Graphics and views in general, remove previous graphic and overview concerns and modify application requirements
For refactoring purporse on the controllers, the views from respective methods needed the methods from concerns. For that, new helpers were created for overview page, graphics and view in a general and the previous concerns were removed. Also, inside the application requirements for helpers and includes were added the new helpers.