Add constant array for the application areas, change verification method parameters and AM/SDR instances names inside the Dashboard controllers
The application controller received a new constant array of the area names of the system. Because of this constant, both AM and SDR dashboard controllers were changed to use this constant instead of the pure string name. Also the Dashboards and manager controller were changed inside the verification method to pass only the controller name and the verification method would fetch the action name inside the application controller.