Create first files(model, controller, database searchs, route, concern and view)
A new forecast feature is been introduced in the application. A forecast is the amount of deals sent by a AM for a client between a certain month. For that, a new Deal table inside DB were created with its model, controller, dabatase searchs, concern file and view. All of the files are the first representation of the Deal structure inside the application.