Change currency format in the contract controller
A bug was found about the contract controller that the update method could not change the contract params because of the contract value currency format. For that, in the controller and the month sales/forecast views were modified to use the right integer values for the create and update methods.